Add Calendar page and refactor Delete
