+++ frontend/src/Calendar.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+function Calendar() {
+  return <div>HI</div>
+}
+
+export default Calendar
